Fixed improper file checking
std::stringstream class of C++ standard library sets badbit flag to signal an error on the stream, while failbit flag is set when there was an error after extraction (no characters to extract or invalid type).
To check validity of input program used to check badbit flag, so it never checked for the validity of the input. This commit changes every check in parseFile function so that it looks at failbit rather than badbit.